# app/api/API_spec.yaml
openapi: 3.0.3
info:
  title: Digitization Toolkit Backend API
  description: |
    API for the Digitization Toolkit - an open-source, modular digitization platform
    for capturing, managing, and organizing archival materials using Raspberry Pi and 64MP ArduCam cameras.
    
    The API provides endpoints for:
    - User authentication and account management
    - Project and collection management
    - Record (document/archival object) metadata management
    - Image capture control and processing
    - Camera device discovery and calibration
    - System monitoring
  version: 1.0.0
  contact:
    name: UCSB AMP Lab
    url: https://github.com/UCSB-AMPLab/digitization-toolkit-backend
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: http://localhost:3000
    description: Production server (frontend proxy)
...
paths:
...
  /cameras/capture:
    post:
      summary: Trigger single image capture
      operationId: trigger_capture
      tags:
        - cameras
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaptureRequest'
      responses:
        '200':
          description: Capture successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaptureResponse'
        '404':
          description: Camera not connected or record not found
  /cameras/capture/dual:
    post:
      summary: Trigger simultaneous dual camera capture
      operationId: trigger_dual_capture
      tags:
        - cameras
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DualCaptureRequest'
      responses:
        '200':
          description: Dual capture successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaptureResponse'
        '404':
          description: Camera not connected or record not found
  /cameras/calibrate:
    post:
      summary: Run autofocus calibration on camera
      operationId: calibrate_camera
      tags:
        - cameras
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalibrationRequest'
      responses:
        '200':
          description: Calibration completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalibrationResponse'
  /cameras/calibrate/white-balance:
    post:
      summary: Calibrate white balance for consistent color reproduction
      operationId: calibrate_white_balance
      tags:
        - cameras
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WhiteBalanceCalibrationRequest'
      responses:
        '200':
          description: White balance calibration completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WhiteBalanceCalibrationResponse'
  /cameras:
    post:
      summary: Create camera settings record
      operationId: create_camera_settings
      tags:
        - cameras
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CameraSettingsCreate'
      responses:
        '200':
          description: Camera settings created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CameraSettingsRead'
        '404':
          description: Record image not found
        '409':
          description: Camera settings already exist

    get:
      summary: List camera settings
      operationId: list_camera_settings
      tags:
        - cameras
      security:
        - BearerAuth: []
      parameters:
        - name: skip
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
      responses:
        '200':
          description: List of camera settings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CameraSettingsRead'
  /cameras/{id}:
    get:
      summary: Get camera settings by ID
      operationId: get_camera_settings
      tags:
        - cameras
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Camera settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CameraSettingsRead'
        '404':
          description: Camera settings not found
  /cameras/settings/{id}:
    put:
      summary: Update camera settings
      operationId: update_camera_settings
      tags:
        - cameras
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CameraSettingsUpdate'
      responses:
        '200':
          description: Camera settings updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CameraSettingsRead'
        '404':
          description: Camera settings not found

    delete:
      summary: Delete camera settings
      operationId: delete_camera_settings
      tags:
        - cameras
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Camera settings deleted
        '404':
          description: Camera settings not found

  # ============================================================================
  # System Endpoints
  # ============================================================================
...
components:
  schemas:
...
    CaptureRequest:
      type: object
      required:
        - project_name
      properties:
        project_name:
          type: string
          example: Ancient Manuscripts
        camera_index:
          type: integer
          default: 0
        resolution:
          type: string
          enum: [low, medium, high]
          default: medium
        include_resolution_in_filename:
          type: boolean
          default: false
        record_id:
          type: integer
          nullable: true
          description: Link to existing record, or create new if None
        record_title:
          type: string
          nullable: true
          description: Used if creating new record
    DualCaptureRequest:
      type: object
      required:
        - project_name
      properties:
        project_name:
          type: string
        resolution:
          type: string
          enum: [low, medium, high]
          default: medium
        include_resolution_in_filename:
          type: boolean
          default: false
        stagger_ms:
          type: integer
          description: Milliseconds between camera triggers
          default: 20
        record_id:
          type: integer
          nullable: true
        record_title:
          type: string
          nullable: true
        sequence:
          type: integer
          nullable: true
          description: Page number/order
    CaptureResponse:
      type: object
      properties:
        success:
          type: boolean
        file_path:
          type: string
          nullable: true
        file_paths:
          type: array
          items:
            type: string
          nullable: true
        record_id:
          type: integer
          nullable: true
        image_ids:
          type: array
          items:
            type: integer
          nullable: true
        timing:
          type: object
          nullable: true
        error:
          type: string
          nullable: true
    CalibrationRequest:
      type: object
      properties:
        camera_index:
          type: integer
          default: 0
        resolution:
          type: string
          enum: [low, medium, high]
          default: high
    CalibrationResponse:
      type: object
      properties:
        success:
          type: boolean
        lens_position:
          type: number
          format: float
          nullable: true
        distance_meters:
          type: number
          format: float
          nullable: true
        af_time:
          type: number
          format: float
          nullable: true
        error:
          type: string
          nullable: true
    WhiteBalanceCalibrationRequest:
      type: object
      properties:
        camera_index:
          type: integer
          default: 0
        resolution:
          type: string
          enum: [low, medium, high]
          default: high
        stabilization_frames:
          type: integer
          default: 30
    WhiteBalanceCalibrationResponse:
      type: object
      properties:
        success:
          type: boolean
        awb_gains:
          type: array
          items:
            type: number
            format: float
          nullable: true
        colour_temperature:
          type: integer
          nullable: true
        converged:
          type: boolean
          nullable: true
        error:
          type: string
          nullable: true
    CameraSettingsCreate:
      type: object
      required:
        - record_image_id
      properties:
        record_image_id:
          type: integer
        camera_model:
          type: string
        iso:
          type: integer
          nullable: true
        aperture:
          type: number
          nullable: true
        focal_length:
          type: number
          nullable: true
        white_balance:
          type: string
          nullable: true
    CameraSettingsUpdate:
      type: object
      properties:
        camera_model:
          type: string
        iso:
          type: integer
          nullable: true
        aperture:
          type: number
          nullable: true
        focal_length:
          type: number
          nullable: true
        white_balance:
          type: string
          nullable: true
    CameraSettingsRead:
      type: object
      properties:
        id:
          type: integer
        record_image_id:
          type: integer
        camera_model:
          type: string
        iso:
          type: integer
          nullable: true
        aperture:
          type: number
          nullable: true
        focal_length:
          type: number
          nullable: true
        white_balance:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/register
...